+++ core/testsuite/src/test/java/com/blazebit/persistence/testsuite/TreatedEntityJoinTest.java
@@ -89,7 +89,9 @@ public class TreatedEntityJoinTest extends AbstractCoreTest {
         root2 = em.createQuery("SELECT e FROM PolymorphicSub1 e WHERE e.sub1Value = 2", PolymorphicSub1.class).getSingleResult();
     }
 
+    // NOTE: EclipseLink pushes uses types used in TREAT expressions into the join group of the base node which is wrong
     @Test
+    @Category({ NoEclipselink.class })
     public void test1() {
         CriteriaBuilder<PolymorphicSub1> cb = cbf.create(em, PolymorphicSub1.class, "root")
                 .innerJoinOn("TREAT(parent AS PolymorphicSub2)", PolymorphicBase.class, "parentAlias")
@@ -98,14 +100,16 @@ public class TreatedEntityJoinTest extends AbstractCoreTest {
 
         String expectedQuery = "SELECT root FROM PolymorphicSub1 root " +
                 "LEFT JOIN root.parent parent_1 " +
-                "JOIN PolymorphicBase parentAlias" + onClause("(TYPE(parent_1) IN (" + PolymorphicSub2.class.getSimpleName() + ") AND parentAlias.id = " + treatRoot("parent_1", PolymorphicSub2.class, "id") + ")");
+                "JOIN PolymorphicBase parentAlias" + onClause("parentAlias.id = " + treatRoot("parent_1", PolymorphicSub2.class, "id"));
         assertEquals(expectedQuery, cb.getQueryString());
         List<PolymorphicSub1> result = cb.getResultList();
         assertEquals(1, result.size());
         assertEquals(root2.getId(), result.get(0).getId());
     }
 
+    // NOTE: EclipseLink pushes uses types used in TREAT expressions into the join group of the base node which is wrong
     @Test
+    @Category({ NoEclipselink.class })
     public void test2() {
         CriteriaBuilder<PolymorphicSub1> cb = cbf.create(em, PolymorphicSub1.class, "root")
                 .innerJoinOn("TREAT(parent AS PolymorphicSub2)", PolymorphicBase.class, "parent")
@@ -114,7 +118,7 @@ public class TreatedEntityJoinTest extends AbstractCoreTest {
 
         String expectedQuery = "SELECT root FROM PolymorphicSub1 root " +
                 "LEFT JOIN root.parent parent_1 " +
-                "JOIN PolymorphicBase parent" + onClause("(TYPE(parent_1) IN (" + PolymorphicSub2.class.getSimpleName() + ") AND parent.id = " + treatRoot("parent_1", PolymorphicSub2.class, "id") + ")");
+                "JOIN PolymorphicBase parent" + onClause("parent.id = " + treatRoot("parent_1", PolymorphicSub2.class, "id"));
         assertEquals(expectedQuery, cb.getQueryString());
         List<PolymorphicSub1> result = cb.getResultList();
         assertEquals(1, result.size());
@@ -130,7 +134,7 @@ public class TreatedEntityJoinTest extends AbstractCoreTest {
         criteria.innerJoinOn("TREAT(p AS PolymorphicSub1).relation1", PolymorphicSub1.class, "r")
                     .on("r").eqExpression("TREAT(p AS PolymorphicSub1).relation1")
                 .end();
-        assertEquals("SELECT p FROM PolymorphicBase p JOIN PolymorphicSub1 r" + onClause("(TYPE(p) IN (PolymorphicSub1) AND r = " + treatRoot("p", PolymorphicSub1.class, "relation1") + ")"), criteria.getQueryString());
+        assertEquals("SELECT p FROM PolymorphicBase p JOIN PolymorphicSub1 r" + onClause("r = " + treatRoot("p", PolymorphicSub1.class, "relation1", true)), criteria.getQueryString());
         criteria.getResultList();
     }
 }