+++ integration/datanucleus-5.1/src/main/java/com/blazebit/persistence/integration/datanucleus/DataNucleus51JpaProvider.java
@@ -516,6 +516,11 @@ public class DataNucleus51JpaProvider implements JpaProvider {
     }
 
     @Override
+    public boolean needsCaseWhenElseBranch() {
+        return false;
+    }
+
+    @Override
     public void setCacheable(Query query) {
         query.setHint("datanucleus.query.results.cached", true);
     }