+++ integration/datanucleus/src/main/java/com/blazebit/persistence/integration/datanucleus/DataNucleusJpaProvider.java
@@ -517,6 +517,11 @@ public class DataNucleusJpaProvider implements JpaProvider {
     }
 
     @Override
+    public boolean needsCaseWhenElseBranch() {
+        return false;
+    }
+
+    @Override
     public void setCacheable(Query query) {
         query.setHint("datanucleus.query.results.cached", true);
     }