+++ integration/eclipselink/src/main/java/com/blazebit/persistence/integration/eclipselink/EclipseLinkJpaProvider.java
@@ -620,6 +620,11 @@ public class EclipseLinkJpaProvider implements JpaProvider {
     }
 
     @Override
+    public boolean needsCaseWhenElseBranch() {
+        return true;
+    }
+
+    @Override
     public void setCacheable(Query query) {
         query.setHint("eclipselink.query-results-cache", true);
     }